+++ cmuxTests/WorkspaceUnitTests.swift
@@ -425,24 +425,12 @@ final class WorkspaceRenameShortcutDefaultsTests: XCTestCase {
             KeyboardShortcutSettings.Action.toggleRightSidebar.label,
             String(localized: "shortcut.toggleRightSidebar.label", defaultValue: "Toggle Right Sidebar")
         )
-        XCTAssertEqual(
-            KeyboardShortcutSettings.Action.openFileExplorerSelection.label,
-            String(localized: "shortcut.openFileExplorerSelection.label", defaultValue: "File Explorer: Open Selection")
-        )
-
         let toggleRightSidebar = KeyboardShortcutSettings.Action.toggleRightSidebar.defaultShortcut
         XCTAssertEqual(toggleRightSidebar.key, "b")
         XCTAssertTrue(toggleRightSidebar.command)
         XCTAssertFalse(toggleRightSidebar.shift)
         XCTAssertTrue(toggleRightSidebar.option)
         XCTAssertFalse(toggleRightSidebar.control)
-
-        let openSelection = KeyboardShortcutSettings.Action.openFileExplorerSelection.defaultShortcut
-        XCTAssertEqual(openSelection.key, "↓")
-        XCTAssertTrue(openSelection.command)
-        XCTAssertFalse(openSelection.shift)
-        XCTAssertFalse(openSelection.option)
-        XCTAssertFalse(openSelection.control)
 
         let focusRightSidebar = KeyboardShortcutSettings.Action.focusRightSidebar.defaultShortcut
         XCTAssertEqual(focusRightSidebar.key, "e")
@@ -513,28 +501,6 @@ final class WorkspaceRenameShortcutDefaultsTests: XCTestCase {
         XCTAssertTrue(shortcut.control)
         XCTAssertTrue(KeyboardShortcutSettings.publicShortcutActions.contains(.markOldestUnreadAndJumpNext))
         XCTAssertTrue(KeyboardShortcutSettings.settingsVisibleActions.contains(.markOldestUnreadAndJumpNext))
-    }
-
-    func testSettingsVisibleShortcutActionsColocateRightSidebarFileExplorerAndFindShortcuts() {
-        let visibleActions = KeyboardShortcutSettings.settingsVisibleActions
-        let expectedActions: [KeyboardShortcutSettings.Action] = [
-            .focusRightSidebar,
-            .toggleRightSidebar,
-            .openFileExplorerSelection,
-            .findInDirectory,
-        ]
-
-        guard let startIndex = visibleActions.firstIndex(of: .focusRightSidebar) else {
-            XCTFail("Toggle Right Sidebar Focus should be visible in keyboard shortcut settings")
-            return
-        }
-
-        let endIndex = startIndex + expectedActions.count
-        guard endIndex <= visibleActions.count else {
-            XCTFail("Expected shortcut settings to include the full right-sidebar shortcut run")
-            return
-        }
-        XCTAssertEqual(Array(visibleActions[startIndex..<endIndex]), expectedActions)
     }
 
     func testMenuItemKeyEquivalentHandlesArrowAndTabKeys() {
